multiple source files support, second step

+++ common_test/32_multiple_files/function.v
@@ -2,5 +2,5 @@
 
 fn average_8b(x i8, y i8) i8 {
 	n := 2
-	return (x + y) / n
+	return i8((x + y) / n)
 }
